consumer-python/requirements.txt
confluent-kafka==2.3.0
psycopg2-binary==2.9.9
redis==5.0.1
requests==2.31.0
structlog==23.2.0
pytest==7.4.4
pytest-mock==3.12.0
